+++ src/util/string-util.spec.ts
